fix: 상수를 constants 폴더로 이동
